pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aws-qwen"
version = "0.1.0"
description = "aws-qwen"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "NguyenSome" }]
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.1.1",
    "memory-profiler>=0.61.0",
    "mlflow>=3.4.0",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "peft>=0.17.1",
    "psutil>=7.1.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "torch>=2.8.0",
    "transformers>=4.56.2",
]

[project.optional-dependencies]
dev = ["ruff", "black", "pytest", "pytest-cov", "pre-commit"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.black]
line-length = 88
target-version = ["py312"]
